requirements.txt
PyHamcrest==2.0.3
requests==2.27.1
streamlit==1.11.0
streamlit_lottie==0.0.3
